Changes Long to Int, ALL

+++ artesanias-01-logic/src/main/java/co/edu/uniandes/csw/artesanias/entities/BaseEntity.java
@@ -19,22 +19,22 @@ import javax.persistence.MappedSuperclass;
 class BaseEntity implements Serializable{
    @Id
    @GeneratedValue(strategy =GenerationType.IDENTITY)
-   private int id;
+   private Long id;
 
     public BaseEntity() {
     }
 
-    public BaseEntity(int id) {
+    public BaseEntity(Long id) {
         this.id = id;
     }
    
    
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
